Refactor tag formatting to draw out shape of output
In the midst of split lines and function calls I found it hard to visualize
what the output would look like given the code producing it. In this patch
we're expanding some of the inline expressions and abstracting a leading newline
in `formatTag` to make the individual formatters look more like what they produce.